add more athlete_info

+++ app_run/models.py
@@ -19,7 +19,7 @@ class Run(models.Model):
 
 class AthleteInfo(models.Model):
     goals = models.TextField(blank=True, default='')
-    weight = models.FloatField(blank=True, null=True)
+    weight = models.IntegerField(blank=True, null=True)
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='athlete_info')
 
     def __str__(self):
